move EntryInfo to file_server.ts
